FU-114 refactor: group detail section
