+++ src/containers/photographer/reservation/section/details/schedule-details.tsx
@@ -1,8 +1,8 @@
 import { useFormContext } from "react-hook-form";
 import { Details } from "reservation-types";
-import SectionLayout from "./section-layout";
-import { sectionStyles } from "./section.css";
-import ScheduleField from "./fields/schedule-field";
+import SectionLayout from "../section-layout";
+import { sectionStyles } from "../section.css";
+import ScheduleField from "../fields/schedule-field";
 
 const ScheduleDetails = () => {
   const { watch } = useFormContext<Details>();
